change: HomeScreenファイル内にあった関数をcomponentに分けました
