+++ app/src/main/java/com/example/tsundokun/ui/home/component/webPageList/webPageCard/GetOgpImageUrl.kt
@@ -1,0 +1,34 @@
+package com.example.tsundokun.ui.home.component.webPageList.webPageCard
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import org.jsoup.Jsoup
+
+/*
+ * htmlからogp画像のurlを取得
+ */
+@Composable
+fun getOgpImageUrl(html: String?): String? {
+    var imageUrl by remember { mutableStateOf<String?>(null) }
+
+    LaunchedEffect(key1 = html) {
+        withContext(Dispatchers.IO) {
+            try {
+                val doc = html?.let { Jsoup.parse(it) }
+                val ogImage = doc?.selectFirst("meta[property=og:image]")
+                if (ogImage != null) {
+                    imageUrl = ogImage.attr("content")
+                }
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
+        }
+    }
+    return imageUrl
+}
